+++ src/main/java/me/jcala/jmooc/service/inter/CrsSer.java
@@ -9,7 +9,7 @@ import java.util.Set;
 
 public interface CrsSer {
 
-    int addCourse(Course course);//添加课程
+    long addCourse(Course course);//添加课程
 
     Set<Chapter> getChapterList(int crsId);//获取某课程的章节列表
